Feat(contact): Them truong so thich cho lien he

+++ app/services/contact.service.js
@@ -12,12 +12,19 @@ class ContactService {
       address: payload.address,
       phone: payload.phone,
       favorite: payload.favorite,
+      hobbies: payload.hobbies,
+      ownerId: payload.ownerId,
+      gender: payload.gender,
+      job: payload.job,
+      birthDate: payload.birthDate,
+      note: payload.note,
     };
 
     // Remove undefined fields
     Object.keys(contact).forEach(
       (key) => contact[key] === undefined && delete contact[key],
     );
+
     return contact;
   }
 
@@ -36,21 +43,25 @@ class ContactService {
     return await cursor.toArray();
   }
 
-  async findByName(name) {
+  async findByName(name, ownerId) {
     return await this.find({
       name: { $regex: new RegExp(name), $options: "i" },
+      // name: { $regex: name, $options: "i" },
+      ownerId: ownerId,
     });
   }
 
-  async findById(id) {
+  async findById(id, ownerId) {
     return await this.Contact.findOne({
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      ownerId: ownerId,
     });
   }
 
-  async update(id, payload) {
+  async update(id, ownerId, payload) {
     const filter = {
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      ownerId: ownerId,
     };
     const update = this.extractConactData(payload);
 
@@ -63,19 +74,20 @@ class ContactService {
     return result;
   }
 
-  async delete(id) {
+  async delete(id, ownerId) {
     const result = await this.Contact.findOneAndDelete({
       _id: ObjectId.isValid(id) ? new ObjectId(id) : null,
+      ownerId: ownerId,
     });
     return result;
   }
 
-  async findFavorite() {
-    return await this.find({ favorite: true });
+  async findFavorite({ ownerId }) {
+    return await this.find({ favorite: true, ownerId });
   }
 
-  async deleteAll() {
-    const result = await this.Contact.deleteMany({});
+  async deleteAll(ownerId) {
+    const result = await this.Contact.deleteMany({ ownerId });
     return result.deletedCount;
   }
 }
